Added Launch Control Homebrew cask.
This is a powerful tool for managing services. The following is taken
from their web site:

"With LaunchControl you see all services and their respective status at
a glance. Invalid services are highlighted and a problem description is
provided. You can enable or disable services with a single click. The
same goes for loading, unloading and ad-hoc starting. A long list of
jobs may be filtered by job name and/or various properties, helping you
to find what you are looking for in an instant."