Updated software setup and restore scripts.
Simplified the setup script so only the bare essentials are
installed/configured. Some duplicaiton is left between setup and
restore in case the backup isn't readily available so the machine
remains immediately usable.

Once the backup is restored, the additional setup scripts will get the
machine ready for permanent use (which was previously done manually).