Updated to Kaleidoscope 3.4.4
Necessary to resolve issues with the `ksdiff` Homebrew cask install
because the App Store app was incompatible with the CLI. Now both the
app and the CLI are in sync. This also means the App Store version of
the app can no longer be used but is now a custom application install.